Feature admin/delete. Closes #15

+++ src/Controller/admin/adminTrickController.php
@@ -3,12 +3,14 @@
 namespace App\Controller\admin;
 
 use App\Entity\Trick;
+use App\Form\TrickType;
 use App\Repository\TrickRepository;
-use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
+use Doctrine\Common\Persistence\ObjectManager;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class AdminTrickController extends AbstractController
 {
@@ -23,24 +25,73 @@ class AdminTrickController extends AbstractController
     }
 
     /**
-     * @Route("/admin/tricks/delete/{id}", name="trick.delete")
+     * @Route("/admin/trick/create", name="admin.trick.new")
+     * @return Symfony\Component\HttpFoundation\Response
      */
-    public function delete(Trick $trick)
+    public function new(Request $request)
     {
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($trick);
-        $entityManager->flush();
+        $trick = new Trick();
 
-        return $this->redirectToRoute('home');
+        $form = $this->createForm(TrickType::class, $trick);
+        $form->handleRequest($request);
+
+        if($form->isSubmitted() && $form->isValid())
+        {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->persist($trick);
+            $entityManager->flush();
+            return $this->redirectToRoute('tricks.index');
+        }
+
+        return $this->render('admin/trick/new.html.twig', [
+            'trick'     =>  $trick,
+            'form'      =>  $form->createView()
+        ]);
     }
 
     /**
-     * @Route("/admin/tricks/edit/{id}", name="trick.edit")
+     * @Route("/admin/trick/{id}", name="admin.trick.edit", methods="GET|POST")
+     * @param Trick $trick
+     * @param Request $request
+     * @return Symfony\Component\HttpFoundation\Response
      */
-    public function edit()
+    public function edit(Trick $trick, Request $request)
     {
+        $form = $this->createForm(TrickType::class, $trick);
+        $form->handleRequest($request);
 
+        if($form->isSubmitted() && $form->isValid())
+        {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->flush();
+            return $this->redirectToRoute('tricks.index');
+        }
+
+
+        return $this->render('admin/trick/edit.html.twig', [
+            'trick'     =>  $trick,
+            'form'      =>  $form->createView()
+        ]);
     }
 
+    /**
+     * @Route("/admin/trick/{id}", name="admin.trick.delete", methods="DELETE")
+     * @param Trick $trick
+     * @return \Symfony\Component\HttpFoundation\Response
+     */
+    public function delete(Trick $trick, Request $request, Trick $tricks)
+    {
+        
 
+        if($this->isCsrfTokenValid('delete' . $trick->getId(), $request->get('_token')))
+        {
+            $entityManager = $this->getDoctrine()->getManager();
+            $entityManager->remove($trick);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('tricks.index');
+        }
+
+        return $this->redirectToRoute('pages/home.html.twig');
+    }
 }
